Restricted Pages

+++ base/views.py
@@ -1,5 +1,7 @@
 from django.shortcuts import render,redirect
 from django.db.models import Q
+from django.http import HttpResponse
+from django.contrib.auth.decorators import login_required
 from .models import Room,Topic
 from django.contrib import messages
 from django.contrib.auth.models import User
@@ -8,6 +10,9 @@ from .forms import RoomForm
 
 
 def LoginPage(request):
+
+    if request.user.is_authenticated:
+        return redirect('home')
     
     if request.method == 'POST':
         username = request.POST.get('username')
@@ -52,6 +57,7 @@ def room(request,pk):
     context = {'room':room}
     return render(request, 'base/room.html',context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -63,9 +69,14 @@ def createRoom(request):
     context = {'form':form}
     return render(request,'base/room_form.html',context)
 
+@login_required(login_url='login')
 def updateRoom(request,pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    
+    if request.user != room.host:
+        return HttpResponse("you are not allowed here !!")
 
     if request.method == 'POST':
         form = RoomForm(request.POST,instance=room)
@@ -76,8 +87,13 @@ def updateRoom(request,pk):
     context = {'form':form}
     return render(request, 'base/room_form.html',context)
 
+@login_required(login_url='login')
 def deleteRoom(request,pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse("you are not allowed here !!")
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
